Added RelocationRecord and relocation list to Section class; Continued implementation of CALL instruction

+++ auxiliary/inc/converters.hpp
@@ -11,9 +11,9 @@ namespace converter
   std::string directive_type_to_string(type::DIRECTIVE_TYPE dir_alias);
   std::string instruction_type_to_string(type::INSTRUCTION_TYPE ins_alias);
 
-  std::array<type::byte, 2> disp_to_byte_arr(int displacement); // returns byte array with 2 elements
-  int get_disp_from_instruction(type::instruction_size instruction_record);
-  int get_negative_val_disp(uint16_t disp_val);
+  std::array<type::byte, 2> disp_to_byte_arr(int16_t displacement); // returns byte array with 2 elements
+  int16_t get_disp_from_instruction(type::instruction_size instruction_record);
+  int16_t get_negative_val_disp(uint16_t disp_val);
 
   type::byte get_upper_half_byte(type::byte single_byte); // TODO: test
   type::byte get_lower_half_byte(type::byte single_byte); // TODO: test
